requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
geoalchemy2==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
celery==5.4.0
redis==5.0.4
rapidfuzz==3.9.3
qrcode[pil]==7.4.2
twilio==9.1.0
pandas==2.2.2
pytest==8.2.0
httpx==0.27.0
python-multipart==0.0.22
